pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-adapter"
version = "0.2.9"
description = "Convert any AI agent into an A2A Protocol server in 3 lines"
readme = "README.md"
license = "Apache-2.0"
authors = [
    {name = "HYBRO AI", email = "info@hybro.ai"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "a2a-sdk[http-server]>=1,<2",
    "grpcio>=1.60.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
n8n = []
crewai = ["crewai>=0.1.0"]
langchain = ["langchain>=0.1.0", "langchain-core>=0.1.0"]
langgraph = ["langgraph>=0.0.1"]
ollama = []
all = [
    "crewai>=0.1.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.1",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/hybroai/a2a-adapter"
Documentation = "https://github.com/hybroai/a2a-adapter#readme"
Repository = "https://github.com/hybroai/a2a-adapter"

[tool.setuptools]
packages = ["a2a_adapter", "a2a_adapter.integrations"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
